user-service/pyproject.toml
[project]
name = "userservice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.7",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "pillow>=11.3.0",
    "psycopg2>=2.9.11",
    "requests>=2.32.5",
    "drf-spectacular>=0.28.0",
    "python-jose>=3.3.0",
    "python-dotenv>=1.0.0",
    "pika>=1.3.2",
]
